requirements.txt
networkx
matplotlib
Pillow
requests
numpy
